Alternative: move ImmutableSpanContext to internal package
